Move ParseError into crate root
